Day 9 part 1 done
New class Task90 made.

MoonComputer refactored to use map instead of vector as a program memory in case jumps over MoonComputer memory are made. With vectors, multiple pushbacks could be needed which means more heap memory needed.

+++ tasks/task71.cpp
@@ -39,7 +39,7 @@ void Task71::execute()
 				int prevAmpIdx = (i + NUM_OF_AMPS - 1) % NUM_OF_AMPS;
 				if (programDone[i])
 					continue;
-				std::vector<int> autoInsertData;
+				std::vector<long long> autoInsertData;
 				if (!result.at(prevAmpIdx).empty()) {
 					parentAmpResult.at(i) = result.at(prevAmpIdx);
 					result.at(prevAmpIdx).clear();
